Created customtab for Terms and reorder.
