+++ app/src/main/java/com/preguardia/app/customtab/CustomTabActivityHelper.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.preguardia.app.checkout;
+package com.preguardia.app.customtab;
 
 import android.app.Activity;
 import android.net.Uri;
