+++ app/src/main/java/com/preguardia/app/customtab/CustomTabsHelper.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.preguardia.app.checkout;
+package com.preguardia.app.customtab;
 
 import android.content.Context;
 import android.content.Intent;
