+++ app/src/main/java/com/preguardia/app/customtab/KeepAliveService.java
@@ -12,23 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.preguardia.app.checkout;
+package com.preguardia.app.customtab;
 
-import android.support.customtabs.CustomTabsClient;
+import android.app.Service;
+import android.content.Intent;
+import android.os.Binder;
+import android.os.IBinder;
 
 /**
- * Callback for events when connecting and disconnecting from Custom Tabs Service.
+ * Empty service used by the custom tab to bind to, raising the application's importance.
  */
-public interface ServiceConnectionCallback {
-    /**
-     * Called when the service is connected.
-     *
-     * @param client a CustomTabsClient
-     */
-    void onServiceConnected(CustomTabsClient client);
+public class KeepAliveService extends Service {
+    private static final Binder sBinder = new Binder();
 
-    /**
-     * Called when the service is disconnected.
-     */
-    void onServiceDisconnected();
+    @Override
+    public IBinder onBind(Intent intent) {
+        return sBinder;
+    }
 }
