+++ app/src/main/res/values/strings.xml
@@ -19,7 +19,7 @@
     <string name="user_register_patient">I\'m patient</string>
     <string name="user_register_medic">I\'m medic</string>
     <string name="user_register_surname">Surname</string>
-    <string name="user_register_email">eMail</string>
+    <string name="user_register_email">Email</string>
     <string name="user_register_date">Birth date</string>
     <string name="user_register_medical">Medical Care</string>
     <string name="user_register_phone">Phone</string>
